feat(middleware): Add auth middleware for route protection, redirection
Implement authentication middleware to safeguard private routes and redirect unauthenticated users. This middleware checks for valid session tokens and redirect users to /login route or login page  or /dashboard route or dashboard page if they are authenticated.

Modify / route to redirect to /login route if user is unauthenticated and if they are authenticated redirct them to /dashboard route.